+++ extension/edr/Tests/EDRExtensionLogicTests/ApplicationControlStoreTests.swift
@@ -21,78 +21,9 @@ import Foundation
 @testable import EDRExtensionLogic
 import XCTest
 
-final class ApplicationControlStoreTests: XCTestCase {
-    // MARK: Helpers
-
-    /// RuleSpec is a named record for the document() helper's `rules` argument.
-    /// A 3-tuple would trip SwiftLint's large_tuple rule (max 2 members); a struct
-    /// also reads better at the call sites and lets us name the field at use.
-    private struct RuleSpec {
-        let type: String
-        let identifier: String
-        let ruleID: String
-    }
-
-    /// makeStore returns a fresh ApplicationControlStore with a temp on-disk
-    /// policy path. The path is unique per call so concurrent / parallelized
-    /// test runs cannot stomp each other's persist files. The teardown block
-    /// removes the file (and any parent directory the persist code created).
-    private func makeStore() -> ApplicationControlStore {
-        let url = FileManager.default.temporaryDirectory
-            .appendingPathComponent("AppControlTests-\(UUID().uuidString)", isDirectory: true)
-            .appendingPathComponent("application-control.json")
-        addTeardownBlock {
-            try? FileManager.default.removeItem(at: url.deletingLastPathComponent())
-        }
-        return ApplicationControlStore(storagePath: url.path)
-    }
-
-    /// waitForFile polls for the file at `path` to exist within `deadline`. The
-    /// store's persistQueue is private, so we cannot synchronize on it directly;
-    /// the file-exists predicate is what the apply→persist→load round-trip test
-    /// needs to gate on before loading the snapshot back from disk. 2s is the
-    /// same budget testStartLazyFillPopulatesCacheEventually uses for the same
-    /// pattern.
-    @discardableResult
-    private func waitForFile(at path: String, deadline: TimeInterval = 2) -> Bool {
-        let stop = Date().addingTimeInterval(deadline)
-        while Date() < stop {
-            if FileManager.default.fileExists(atPath: path) {
-                return true
-            }
-            Thread.sleep(forTimeInterval: 0.01)
-        }
-        return FileManager.default.fileExists(atPath: path)
-    }
-
-    /// document builds an ApplicationControlDocument as on-the-wire JSON bytes. `epoch` is optional so existing callers
-    /// produce a payload with no `policy_epoch` key (decoded as 0, the legacy / pre-#322 path); the regression tests pass an
-    /// explicit epoch to exercise the restore-resync gate.
-    private func document(policyID: Int64, version: Int64, rules: [RuleSpec], epoch: Int64? = nil) -> Data {
-        let ruleObjects = rules.map { rule in
-            """
-            {
-              "rule_id": "\(rule.ruleID)",
-              "rule_type": "\(rule.type)",
-              "identifier": "\(rule.identifier)",
-              "action": "BLOCK",
-              "enforcement": "PROTECT",
-              "severity": "high"
-            }
-            """
-        }
-        let joined = ruleObjects.joined(separator: ",")
-        let epochLine = epoch.map { "  \"policy_epoch\": \($0),\n" } ?? ""
-        let json = """
-        {
-          "policy_id": \(policyID),
-          "policy_version": \(version),
-        \(epochLine)  "rules": [\(joined)]
-        }
-        """
-        return Data(json.utf8)
-    }
-
+// Gate-behaviour + persistence tests for ApplicationControlStore. Shared fixtures (makeStore / document / RuleSpec) live on
+// AppControlStoreTestCase; the #322 epoch / re-sync tests live in ApplicationControlStoreResyncTests.
+final class ApplicationControlStoreTests: AppControlStoreTestCase {
     // MARK: ApplicationControlDocument decoder
 
     func testDocumentDecodesValidJSON() throws {
@@ -230,134 +161,6 @@ final class ApplicationControlStoreTests: XCTestCase {
         // rule list does not include it. Pinning this prevents a future bug
         // where apply() accidentally merges rather than replacing.
         XCTAssertNil(snapshot.binaryRules["baseline"], "newer doc must replace, not merge")
-    }
-
-    // MARK: apply: epoch-axis re-sync after a server policy-version regression (#322)
-
-    // A server DB restore/reset regresses policy_version below what a host has persisted. The version-only gate would freeze
-    // enforcement forever; the epoch axis (the policy's updated_at in microseconds) keeps advancing across the restore, so the
-    // host re-syncs. This pins: baseline at a high version, then a LOWER version whose epoch advanced -> accepted, ruleset
-    // replaced wholesale (the freeze is gone).
-    // swiftlint:disable:next line_length
-    func test_spec_extension_application_control_snapshot_is_the_source_of_truth_for_decisions_a_version_regression_with_a_newer_epoch_re_syncs_instead_of_freezing() {
-        let store = makeStore()
-
-        // Pre-restore baseline: high version, earlier epoch.
-        store.apply(rawJSON: document(
-            policyID: 1, version: 25,
-            rules: [RuleSpec(type: "BINARY", identifier: "pre-restore", ruleID: "pre")],
-            epoch: 1_000
-        ))
-        XCTAssertEqual(store.currentSnapshot().policyVersion, 25)
-
-        // Post-restore push: version regressed to 2, but the operator's next mutation stamped a fresh (larger) updated_at.
-        store.apply(rawJSON: document(
-            policyID: 1, version: 2,
-            rules: [RuleSpec(type: "BINARY", identifier: "post-restore", ruleID: "post")],
-            epoch: 2_000
-        ))
-        let snapshot = store.currentSnapshot()
-        XCTAssertEqual(snapshot.policyVersion, 2, "the lower post-restore version is accepted on the epoch axis")
-        XCTAssertEqual(snapshot.policyEpoch, 2_000)
-        XCTAssertNotNil(snapshot.binaryRules["post-restore"], "the post-restore ruleset now enforces")
-        XCTAssertNil(snapshot.binaryRules["pre-restore"], "the stale ruleset is gone; enforcement is no longer frozen")
-    }
-
-    // The gate rejects only when BOTH axes are <= current. A snapshot that is older on version AND epoch is a genuine
-    // duplicate / out-of-order replay (e.g. the newest-first command batch delivering an older entry after a newer one) and
-    // must not regress the active ruleset.
-    func testApplyRejectsSnapshotOlderOnBothAxes() {
-        let store = makeStore()
-        store.apply(rawJSON: document(
-            policyID: 1, version: 25,
-            rules: [RuleSpec(type: "BINARY", identifier: "current", ruleID: "cur")],
-            epoch: 2_000
-        ))
-        // Lower version AND lower epoch -> rejected.
-        store.apply(rawJSON: document(
-            policyID: 1, version: 2,
-            rules: [RuleSpec(type: "BINARY", identifier: "older", ruleID: "old")],
-            epoch: 1_000
-        ))
-        let snapshot = store.currentSnapshot()
-        XCTAssertEqual(snapshot.policyVersion, 25, "older-on-both-axes apply must not regress the ruleset")
-        XCTAssertEqual(snapshot.policyEpoch, 2_000)
-        XCTAssertNil(snapshot.binaryRules["older"])
-        XCTAssertNotNil(snapshot.binaryRules["current"])
-    }
-
-    // Backward-compat: a pre-#322 server (and any snapshot persisted before the field existed) carries no policy_epoch, which
-    // decodes to 0. With both epochs at 0 the gate falls back to version-only behaviour: a regressed version is rejected, a
-    // forward version is accepted. This guarantees the change is inert until the server also emits epochs.
-    func testApplyWithoutEpochFallsBackToVersionOnlyGate() {
-        let store = makeStore()
-        store.apply(rawJSON: document(
-            policyID: 1, version: 25,
-            rules: [RuleSpec(type: "BINARY", identifier: "baseline", ruleID: "base")]
-        ))
-        XCTAssertEqual(store.currentSnapshot().policyEpoch, 0, "absent policy_epoch decodes to the 0 sentinel")
-
-        // Lower version, no epoch (both 0) -> rejected, exactly as before the fix.
-        store.apply(rawJSON: document(
-            policyID: 1, version: 2,
-            rules: [RuleSpec(type: "BINARY", identifier: "regressed", ruleID: "reg")]
-        ))
-        XCTAssertEqual(store.currentSnapshot().policyVersion, 25, "no-epoch regression stays gated by version")
-        XCTAssertNil(store.currentSnapshot().binaryRules["regressed"])
-
-        // Higher version, no epoch -> accepted via the version axis.
-        store.apply(rawJSON: document(
-            policyID: 1, version: 26,
-            rules: [RuleSpec(type: "BINARY", identifier: "forward", ruleID: "fwd")]
-        ))
-        XCTAssertEqual(store.currentSnapshot().policyVersion, 26)
-        XCTAssertNotNil(store.currentSnapshot().binaryRules["forward"])
-    }
-
-    // swiftlint:disable:next line_length
-    func test_spec_extension_application_control_application_control_re_sync_event_a_version_regression_emits_a_re_sync_event() {
-        let store = makeStore()
-        var reported: ApplicationControlResyncPayload?
-        store.resyncReporter = { reported = $0 }
-
-        store.apply(rawJSON: document(
-            policyID: 1, version: 25,
-            rules: [RuleSpec(type: "BINARY", identifier: "pre", ruleID: "pre")],
-            epoch: 1_000
-        ))
-        store.apply(rawJSON: document(
-            policyID: 1, version: 2,
-            rules: [RuleSpec(type: "BINARY", identifier: "post", ruleID: "post")],
-            epoch: 2_000
-        ))
-
-        let payload = try? XCTUnwrap(reported)
-        XCTAssertEqual(payload?.policyID, 1)
-        XCTAssertEqual(payload?.previousVersion, 25)
-        XCTAssertEqual(payload?.newVersion, 2)
-        XCTAssertEqual(payload?.previousEpoch, 1_000)
-        XCTAssertEqual(payload?.newEpoch, 2_000)
-        XCTAssertEqual(payload?.reason, "version_regression")
-    }
-
-    // swiftlint:disable:next line_length
-    func test_spec_extension_application_control_application_control_re_sync_event_a_normal_forward_apply_emits_no_re_sync_event() {
-        let store = makeStore()
-        var fired = false
-        store.resyncReporter = { _ in fired = true }
-
-        // First apply (cold start) and a normal forward apply: version advances, epoch advances. Neither is a regression.
-        store.apply(rawJSON: document(
-            policyID: 1, version: 1,
-            rules: [RuleSpec(type: "BINARY", identifier: "v1", ruleID: "r1")],
-            epoch: 1_000
-        ))
-        store.apply(rawJSON: document(
-            policyID: 1, version: 2,
-            rules: [RuleSpec(type: "BINARY", identifier: "v2", ruleID: "r2")],
-            epoch: 2_000
-        ))
-        XCTAssertFalse(fired, "a forward apply (version advancing) must not emit a re-sync event")
     }
 
     // MARK: apply: atomic in-memory swap to the newer snapshot
